requirements.txt
bpy==4.0.0
Imath==0.0.2
KNN_CUDA==0.2
mathutils==3.3.0
matplotlib==3.5.3
numpy==1.21.6
open3d==0.17.0
opencv_python==4.8.1.78
OpenEXR==1.3.9
Pillow==10.2.0
pytorch_msssim==1.0.0
pyvista==0.43.2
scikit_learn==1.0.2
setuptools==58.2.0
tensorboardX==2.6.2.2
termcolor==2.4.0
torch==1.7.0
torchvision==0.8.0
tqdm==4.66.1
